+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] - 2024-11-13
+
+### Added
+
+-   Added the `circular-wheel` board style.
+-   Added support for different enter/exit annotation shapes.
+
 ## [1.5.1] - 2024-10-24
 
 ### Changed